requirements.txt
pyyaml
ffmpeg-python